common file created for phd and msc = just thesis

+++ controllers/user.controller.js
@@ -10,14 +10,14 @@ userCtrl.login = async (req, res, next) => {
         let data = req.body;
         console.log("data:", data)
         var user = await User.findOne(req.body)
-        if(user){
-            var apiRes = {status: true,message:'You are successfully logged in'}
-        }else{
-            var apiRes = {status: false,message:'Username or Password incorrect'}
+        if (user) {
+            var apiRes = { status: true, message: 'You are successfully logged in' }
+        } else {
+            var apiRes = { status: false, message: 'Username or Password incorrect' }
         }
         console.log(user)
     } catch (error) {
-        return res.status(500).send({status : false,message : 'Something went worng'})
+        return res.status(500).send({ status: false, message: 'Something went worng' })
     }
     return res.status(200).send(apiRes)
 }
@@ -31,17 +31,17 @@ userCtrl.addUser = async (req, res, next) => {
         const { username } = req.body;
         const existingUser = await User.findOne({ username });
         if (existingUser) {
-            return res.status(200).send({ status : false, message: 'Username already exists' });
+            return res.status(200).send({ status: false, message: 'Username already exists' });
         }
 
         var user = await User.create(data)
-        if(user){
-            var apiRes = {status : true,message : 'User added successfully'}
-        }else{
-            var apiRes = {status : false,message : 'Something went worng'}
+        if (user) {
+            var apiRes = { status: true, message: 'User added successfully' }
+        } else {
+            var apiRes = { status: false, message: 'Something went worng' }
         }
     } catch (error) {
-        return res.status(500).send({status : false,message : 'Something went worng'})
+        return res.status(500).send({ status: false, message: 'Something went worng' })
     }
     return res.status(200).send(apiRes)
 }
@@ -57,7 +57,7 @@ userCtrl.getUser = async (req, res, next) => {
 
 userCtrl.getUserById = async (req, res, next) => {
     try {
-        const {_id} = req.query
+        const { _id } = req.query
         console.log(_id)
         var user = await User.findById(_id)
     } catch (error) {
@@ -67,24 +67,24 @@ userCtrl.getUserById = async (req, res, next) => {
 }
 
 userCtrl.editUser = async (req, res, next) => {
-    const {_id} = req.query
+    const { _id } = req.query
     const { username, password } = req.body;
     try {
         const { username } = req.body;
-        
+
         // const existingUser = await User.findOne({ username });
         // if (existingUser) {
         //     return res.status(400).send({ status : false, message: 'Username already exists' });
         // }
         var user = await User.findByIdAndUpdate(_id, { password })
-        if(user){
-            var resData = {status : true, message : 'User updated successfully.'}
-        }else{
-            var resData = {status : false, message : 'User not found.'}
+        if (user) {
+            var resData = { status: true, message: 'User updated successfully.' }
+        } else {
+            var resData = { status: false, message: 'User not found.' }
         }
-        
+
     } catch (error) {
-        return res.status(500).send({status : false, message : 'something went worng'})
+        return res.status(500).send({ status: false, message: 'something went worng' })
     }
     return res.status(200).send(resData)
 }
@@ -92,14 +92,20 @@ userCtrl.editUser = async (req, res, next) => {
 userCtrl.deletUser = async (req, res, next) => {
     const { _id } = req.query;
     try {
-        var user = await User.findByIdAndDelete(_id)
-        if(user){
-            var apiRes = {status:true,message:'User deleted successfully'}
-        }else{
-            var apiRes = {status:false, message:'User not found'}
+        var totalUser = await User.find();
+        if (totalUser.length == 1) {
+            var apiRes = { status: false, message: "Atlease one user is compulsory, you can't delete this user." }
+        } else {
+            var user = await User.findByIdAndDelete(_id)
+            if (user) {
+                var apiRes = { status: true, message: 'User deleted successfully' }
+            } else {
+                var apiRes = { status: false, message: 'User not found' }
+            }
         }
+
     } catch (error) {
-        return res.status(500).send({status:false,message:'Something went wrong.'})
+        return res.status(500).send({ status: false, message: 'Something went wrong.' })
     }
     return res.status(200).send(apiRes)
 }
